requirements.txt
anyio==4.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Uploads==0.2.1
fsspec==2024.2.0
gradio_client==0.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pillow==10.2.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
websockets==11.0.3
Werkzeug==3.0.1
gunicorn
